Added full robot state to the robot status publisher and made everything into float64 aka doubles instead of float32.

+++ iam_robolib/include/iam_robolib/definitions.h
@@ -2,7 +2,7 @@
 #define IAM_ROBOLIB_DEFINITIONS_H_
 
 // SharedBuffer type to share memory (Change size later)
-typedef float* SharedBuffer;
+typedef double* SharedBuffer;
 
 // Enum for Robot Types
 enum class RobotType {
